+++ inst/resources/init.R
@@ -178,7 +178,7 @@ local({
     ## an 'installed from source' version
 
     ## -- InstallAgent -- ##
-    installAgent <- 'InstallAgent: packrat 0.4.5'
+    installAgent <- 'InstallAgent: packrat 0.4.5-1'
 
     ## -- InstallSource -- ##
     installSource <- 'InstallSource: source'