+++ NEWS.md
@@ -1,8 +1,11 @@
 # Packrat 0.4.4 (Unreleased)
 
+- Simple functions for interacting with the set of available repositories
+  have been added. See `?repository-management` for more details.
+  
 - Facilities for interacting with local CRAN-like repositories have been added.
   This feature will eventually supersede the functionality offered by packrat's
-  'ad-hoc' local repositories. The functions `packrat::create_repo()` and
+  'ad-hoc' local repositories. The functions `packrat::repos_create()` and
   `packrat::upload_package()` are the two main API functions currently exported
   for uploading (source) packages to a local CRAN-like repository. By using a
   local CRAN-like repository, pre-existing workflows using `library()` and